requirements.txt
tensorflow==2.20.0        # or compatible version installed on your system
numpy
pillow
scikit-learn
matplotlib
streamlit
ultralytics
opencv-python
